% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bandwidth.R
\name{bandwidth}
\alias{bandwidth}
\alias{bw.scott}
\alias{bw.silv86}
\title{Bandwidth Selectors for Multivariate Kernel Density Estimation}
\usage{
bw.silv86(x)

bw.scott(x)
}
\arguments{
\item{x}{numeric vector.}
}
\description{
Bandwidth selectors for Gaussian kernels
}
\references{
Silverman, B. W. (1986). Density estimation for statistics and data analysis. Chapman and Hall/CRC.

Wand, M. P. and Jones, M. C. (1995). Kernel Smoothing. Chapman and Hall/CRC.

Scott, D. W. (1992). Multivariate density estimation: theory, practice,
and visualization. John Wiley & Sons.
}
\seealso{
\code{\link[stats]{bw.nrd}}
}

