% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bandwidth.R
\name{bw.scott}
\alias{bw.scott}
\alias{bw.silv}
\title{Bandwidth Selector for Multivariate Kernel Density Estimation}
\usage{
bw.scott(x)

bw.silv(x)
}
\arguments{
\item{x}{numeric matrix or data.frame.}
}
\description{
Rule of thumb bandwidth selectors for Gaussian kernels as described by
Scott (1992) and Silverman (1986).
}
\details{
Scott's (1992) rule is defined as

\deqn{
H = \left(n^{-1/(k+4)})^2 \mathrm{diag}(S)
}{
H = [n^(-1/(k+4))]^2 * diag(S)
}

Silverman's (1986) rule is defined as

\deqn{
H = \left(\left(\frac{4}{(k+2)n}\right)^{1/(k+4)}\right)^2 \mathrm{diag}(S)
}{
H = [(4/((k+2)*n))^(1/(k+4))]^2 * diag(S)
}

where \eqn{k} is number of variables, \eqn{n} is sampel size and \eqn{S}
is an empirical  covariance matrix.
}
\references{
Silverman, B. W. (1986). Density estimation for statistics and data analysis. Chapman and Hall/CRC.

Wand, M. P. and Jones, M. C. (1995). Kernel Smoothing. Chapman and Hall/CRC.

Scott, D. W. (1992). Multivariate density estimation: theory, practice,
and visualization. John Wiley & Sons.
}
\seealso{
\code{\link[stats]{bandwidth}}
}
