% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/multivar-pkd.R
\name{dmvpkd}
\alias{dmvpkd}
\alias{rmvpkd}
\title{Multivariate product kernel density}
\usage{
dmvpkd(x, y, bw = sqrt(diag(bw.silv(y))), weights = NULL,
  kernel = c("gaussian", "epanechnikov", "rectangular", "triangular",
  "biweight", "triweight", "cosine", "optcosine"), adjust = 1,
  log.prob = FALSE)

rmvpkd(n, y, bw = sqrt(diag(bw.silv(y))), weights = NULL,
  kernel = c("gaussian", "epanechnikov", "rectangular", "triangular",
  "biweight", "triweight", "cosine", "optcosine"), adjust = 1)
}
\arguments{
\item{x}{numeric vector; kernel density is evaluated on those values.}

\item{y}{numeric vector; kernel density is estimated using those values.}

\item{bw}{numeric matrix}

\item{weights}{numeric vector}

\item{kernel}{a character string giving the smoothing kernel to be used.
This must partially match one of "gaussian", "rectangular",
"triangular", "epanechnikov", "biweight", "cosine" or
"optcosine", with default "gaussian", and may be abbreviated
to a unique prefix (single letter).}

\item{adjust}{scalar; the bandwidth used is actually \code{adjust*bw}.
This makes it easy to specify values like 'half the default'
bandwidth.}

\item{log.prob}{logical; if \code{TRUE}, probabilities p are given as log(p).}

\item{n}{number of observations. If length(n) > 1,
the length is taken to be the number required.}
}
\description{
Multivariate product kernel density
}
\details{
Multivariate product kernel density estimator is defined as a product of univariate kernels

\deqn{
\hat{f_H}(x_1,\dots,x_n) = \sum_{i=1}^n w_i \prod_{j=1}^k
K_{h_j} \left( \frac{x_i - y_{ij}}{h_j} \right)
}{
f(x) = sum[i](w[i] * prod[j]( Khj((x[i]-y[i,j])/h[j]) ))
}

where \eqn{w} is a vector of weights such that \eqn{\sum_i w_i = 1}{sum(w) = 1}
(by default \eqn{w_i=1/n}{w[i]=1/n} for all \eqn{i}), \eqn{K_{h_j}}{Khj} is
kernel \eqn{K} parametrized by bandwidth \eqn{h_j} and \eqn{\boldsymbol{y}}{y}
is a matrix of data points used for estimating the kernel density.
}
\references{
Silverman, B. W. (1986). Density estimation for statistics and data analysis.
Chapman and Hall/CRC.

Wand, M. P. and Jones, M. C. (1995). Kernel Smoothing. Chapman and Hall/CRC.

Scott, D. W. (1992). Multivariate density estimation: theory, practice,
and visualization. John Wiley & Sons.
}
\seealso{
\code{\link{kernelboot}}
}
