% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/univar-kde.R
\name{duvkd}
\alias{duvkd}
\alias{ruvkd}
\title{Univariate kernel density}
\usage{
duvkd(y, x, bw = bw.nrd0(x), weights = rep(1, length(x)),
  kernel = c("gaussian", "epanechnikov", "rectangular", "triangular",
  "biweight", "triweight", "cosine", "optcosine"), log.prob = FALSE)

ruvkd(n, x, bw = bw.nrd0(x), weights = rep(1, length(x)),
  kernel = c("gaussian", "epanechnikov", "rectangular", "triangular",
  "biweight", "triweight", "cosine", "optcosine"), preserve.var = FALSE)
}
\arguments{
\item{y, x}{numeric vector.}

\item{bw}{numeric value.}

\item{weights}{numeric vector.}

\item{kernel}{a character string giving the smoothing kernel to be used.
This must partially match one of "gaussian", "rectangular",
"triangular", "epanechnikov", "biweight", "cosine" or
"optcosine", with default "gaussian", and may be abbreviated
to a unique prefix (single letter).}

\item{log.prob}{logical; if TRUE, probabilities p are given as log(p).}

\item{n}{number of observations. If length(n) > 1,
the length is taken to be the number required.}

\item{preserve.var}{logical;}
}
\description{
Univariate kernel density
}
