% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/univar-kde.R
\name{duvkde}
\alias{duvkde}
\alias{ruvkde}
\title{Univariate kernel density}
\usage{
duvkde(y, x, bw = bw.nrd0(x), weights = rep(1, length(x)),
  kernel = "gaussian", log.prob = FALSE)

ruvkde(n, x, bw = bw.nrd0(x), weights = rep(1, length(x)),
  kernel = "gaussian", preserve.var = FALSE)
}
\arguments{
\item{y, x}{numeric vector.}

\item{bw}{numeric value.}

\item{weights}{numeric vector.}

\item{log.prob}{logical; if TRUE, probabilities p are given as log(p).}

\item{n}{number of observations. If length(n) > 1,
the length is taken to be the number required.}

\item{preserve.var}{logical;}
}
\description{
Univariate kernel density
}
