% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/kernelboot-methods.R
\name{kernelboot-class}
\alias{kernelboot-class}
\alias{summary.kernelboot}
\alias{getSamples}
\alias{as.data.frame.kernelboot}
\title{kernelboot class object}
\usage{
\method{summary}{kernelboot}(object, quantiles = c(0.025, 0.5, 0.975), ...)

getSamples(x, simplify = TRUE)

\method{as.data.frame}{kernelboot}(x, ...)
}
\arguments{
\item{quantiles}{returned quantiles (see \code{\link{quantile}}).}

\item{\dots}{further arguments passed to or from other methods.}

\item{x, object}{\code{kernelboot} class object.}

\item{simplify}{if \code{TRUE} object is simplified to data.frame.}
}
\description{
kernelboot class object
}
\details{
Object of class \code{"kernelboot"}, is a list with components including

\tabular{ll}{
\code{orig.stat}          \tab  estimates from \code{statistic} on the original data, \cr
\code{boot.samples}       \tab  samples drawn, \cr
\code{call}               \tab  function call, \cr
\code{statistic}          \tab  actual \code{statistic} function that was used, \cr
\code{orig.data}          \tab  original data used for bootstrapping, \cr
\code{variables}          \tab  used variables: it is \code{NULL} for univariate data and
                                for multivaiate data it contains two lists of \code{smoothed}
                                and \code{ignored} variables (names or column indexes) during
                                the smoothing phase. \cr
\code{type}               \tab  type of kernel density that was used, \cr
\code{param}              \tab  list of parameters that were used.
}

\code{param} section contains:

\tabular{ll}{
\code{R}                  \tab  number of bootstrap iterations, \cr
\code{bw}                 \tab  the bandwidth that was used, \cr
\code{weights}            \tab  vector of the weights that were applied, \cr
\code{kernel}             \tab  name of the kernel that was used, \cr
\code{shrinked}           \tab  value of the \code{shrinked} parameter, \cr
\code{parallel}           \tab  indicates if parallel computation was used.
}
}
\seealso{
\code{\link{kernelboot}}
}
