% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{KernelRNG}
\alias{KernelRNG}
\alias{rbiweight}
\alias{rcosine}
\alias{rempan}
\alias{roptcos}
\alias{rrect}
\alias{rtriang}
\alias{rtriweight}
\title{Random generation from kernels}
\usage{
rempan(n)

rcosine(n)

roptcos(n)

rtriang(n)

rrect(n)

rbiweight(n)

rtriweight(n)
}
\arguments{
\item{n}{number of observations.}
}
\description{
Random generation from kernels
}

