% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rkde.R
\name{rkde}
\alias{rkde}
\title{Draw samples from kernel density with Gaussian kernel}
\usage{
rkde(n, x, bw = "nrd0", method = c("default", "smoothed"), adjust = 1,
  weights = NULL, na.rm = FALSE)
}
\arguments{
\item{n}{number of observations to be drawn. If \code{length(n) > 1}, the length is taken to be the number required.
If \code{n} is not provided, it defaults to \code{length(x)}.}

\item{x}{the data from which the estimate is to be computed. \emph{Warning:} infinite values are omiitted.}

\item{bw}{the smoothing bandwidth to be used. The kernels are scaled such that this is
the standard deviation of the smoothing kernel. (Note this differs from the
reference books cited below, and from S-PLUS.)
\code{bw} can also be a character string giving a rule to choose the bandwidth. See
\code{\link[stats]{bw.nrd}}. The default, \code{"nrd0"}, has remained the default for historical
and compatibility reasons, rather than as a general recommendation, where e.g., \code{"SJ"}
would rather fit, see also Venables and Ripley (2002).}

\item{method}{algorithm to be used for sampling as described by Silverman (2000), \code{"default"}
is the general algorithm and \code{"smoothed"} is the \emph{Step 3'} variant (p. 143).}

\item{adjust}{the bandwidth used is actually \code{adjust*bw}. This makes it easy to specify values like
‘half the default’ bandwidth.}

\item{weights}{numeric vector of non-negative observation weights, hence of same length as \code{x}.
The default NULL is equivalent to \code{weights = rep(1/nx, nx)} where \code{nx}
is the length of (the finite entries of) \code{x[]}.}

\item{na.rm}{logical; if \code{TRUE}, missing values are removed from \code{x}.
If \code{FALSE} any missing values cause an error.}
}
\description{
Draw samples from kernel density with Gaussian kernel
}
\examples{

hist(rkde(1e5, mtcars$disp), 100, freq = FALSE)
lines(density(mtcars$disp), col = "red")
rug(mtcars$disp, lwd = 2)

}
\references{
Silverman, B.W. (2000). Density estimation for statistics and data analysis.
CRC/Chapman & Hall. (pp. 142-144)

Venables, W.N. and Ripley, B.D. (2002) Modern Applied Statistics with S.
New York: Springer.
}
\seealso{
\code{\link[stats]{density}}
}

