% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rkernel.R
\name{rkernel}
\alias{rkernel}
\title{Draw samples from kernel density}
\usage{
rkernel(n, x, bw = "nrd0", kernel = c("epanechnikov", "gaussian",
  "rectangular", "triangular", "biweight", "triweight", "cosine", "optcosine"),
  preserve.var = TRUE, adjust = 1, weights = NULL, na.rm = FALSE)
}
\arguments{
\item{n}{number of observations to be drawn. If \code{length(n) > 1}, the length is taken to be the number required.
If \code{n} is not provided, it defaults to \code{length(x)}.}

\item{x}{the data from which the estimate is to be computed. \emph{Warning:} infinite values are omiitted.
If not provided, draws samples from given \code{kernel} and uses \code{bw} only if
it was provided as numeric value.}

\item{bw}{the smoothing bandwidth to be used. The kernels are scaled such that this is
the standard deviation of the smoothing kernel. (Note this differs from the
reference books cited below, and from S-PLUS.)
\code{bw} can also be a character string giving a rule to choose the bandwidth. See
\code{\link[stats]{bw.nrd}}. The default, \code{"nrd0"}, has remained the default for historical
and compatibility reasons, rather than as a general recommendation, where e.g., \code{"SJ"}
would rather fit, see also Venables and Ripley (2002).}

\item{kernel}{a character string giving the smoothing kernel to be used.}

\item{preserve.var}{logical, if \code{TRUE}, then the bootstrap samples preserve sample variance.}

\item{adjust}{the bandwidth used is actually \code{adjust*bw}. This makes it easy to specify values like
'half the default' bandwidth.}

\item{weights}{numeric vector of non-negative observation weights, hence of same length as \code{x}.
The default NULL is equivalent to \code{weights = rep(1/nx, nx)} where \code{nx}
is the length of (the finite entries of) \code{x[]}.}

\item{na.rm}{logical; if \code{TRUE}, missing values are removed from \code{x}.
If \code{FALSE} any missing values cause an error.}
}
\description{
Draw samples from kernel density
}
\examples{

hist(rkernel(1e5, mtcars$disp), 100, freq = FALSE)
lines(density(mtcars$disp), col = "red")
rug(mtcars$disp, lwd = 2)

}
\references{
Silverman, B. W. (1986). Density estimation for statistics and data analysis.
Chapman and Hall/CRC.

Wand, M. P. and Jones, M. C. (1995). Kernel Smoothing. Chapman and Hall/CRC.

Scott, D. W. (1992). Multivariate density estimation: theory, practice,
and visualization. John Wiley & Sons.

Venables, W. N. and Ripley, B. D. (2002) Modern Applied Statistics with S.
New York: Springer.
}
\seealso{
\code{\link[stats]{density}}
}

