% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/kernelboot-methods.R
\name{summary.kernelboot}
\alias{summary.kernelboot}
\title{Summarize the result of kernelboot}
\usage{
\method{summary}{kernelboot}(object, probs = c(0.025, 0.5, 0.975), ..., na.rm = FALSE)
}
\arguments{
\item{object}{\code{kernelboot} class object.}

\item{probs}{quantiles returned by \code{summary} (see \code{\link{quantile}}).}

\item{\dots}{further arguments passed to or from other methods.}

\item{na.rm}{a logical value indicating whether \code{NA} values should be
stripped before the computation proceeds.}
}
\description{
Summarize the result of kernelboot
}
