% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bandwidth.R
\name{bandwidth}
\alias{bandwidth}
\alias{bw.scott}
\alias{bw.silv86}
\title{Bandwidth Selectors for Multivariate Kernel Density Estimation}
\usage{
bw.silv86(x)

bw.scott(x)
}
\arguments{
\item{x}{numeric vector.}
}
\description{
Rule of thumb bandwidth selectors for Gaussian kernels as described by
Silverman (1986) and Scott (1992).
}
\details{
Scott's rule is defined as

\deqn{
h_i = n^{-1/(d+4)} \sigma_i
}{
h[i] = n^(-1/(d+4)) * \sigma[i]
}

Silverman's rule is defined as

\deqn{
h_i = \left(\frac{4}{d+2}\right)^{1/(d+4)} n^{-1/(d+4)} \sigma_i
}{
h[i] = (4/(d+2))^(1/(d+4)) * n^(-1/(d+4)) * \sigma[i]
}

where \eqn{d} is number of variables and \eqn{n} is sampel size.
}
\references{
Silverman, B. W. (1986). Density estimation for statistics and data analysis. Chapman and Hall/CRC.

Wand, M. P. and Jones, M. C. (1995). Kernel Smoothing. Chapman and Hall/CRC.

Scott, D. W. (1992). Multivariate density estimation: theory, practice,
and visualization. John Wiley & Sons.
}
\seealso{
\code{\link[stats]{bw.nrd}}
}

