% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bandwidth.R
\name{bw.scott}
\alias{bw.scott}
\title{Bandwidth Selector for Multivariate Kernel Density Estimation}
\usage{
bw.scott(x)
}
\arguments{
\item{x}{numeric vector.}
}
\description{
Rule of thumb bandwidth selector for Gaussian kernels as described by
Scott (1992).
}
\details{
Scott's rule is defined as

\deqn{
h_i = n^{-1/(d+4)} \hat\Sigma
}{
h[i] = n^(-1/(d+4)) * \Sigma
}

where \eqn{d} is number of variables and \eqn{n} is sampel size.
}
\references{
Silverman, B. W. (1986). Density estimation for statistics and data analysis. Chapman and Hall/CRC.

Wand, M. P. and Jones, M. C. (1995). Kernel Smoothing. Chapman and Hall/CRC.

Scott, D. W. (1992). Multivariate density estimation: theory, practice,
and visualization. John Wiley & Sons.
}
\seealso{
\code{\link[stats]{bw.nrd}}
}

