% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/multivar-kde.R
\name{dmvkd}
\alias{dmvkd}
\alias{rmvkd}
\title{Multivariate kernel density}
\usage{
dmvkd(y, x, bw = bw.scott(x), weights = rep(1, nrow(x)), log.prob = FALSE)

rmvkd(n, x, bw = bw.scott(x), weights = rep(1, nrow(x)),
  preserve.var = FALSE)
}
\arguments{
\item{y, x}{numeric matrix.}

\item{bw}{numeric matrix}

\item{weights}{numeric vector}

\item{log.prob}{logical; if TRUE, probabilities p are given as log(p).}

\item{n}{number of observations. If length(n) > 1,
the length is taken to be the number required.}

\item{preserve.var}{logical;}
}
\description{
Multivariate kernel density
}
