% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mvn-arma.R
\name{dmvn}
\alias{dmvn}
\alias{rmvn}
\title{Multivariate normal distribution}
\usage{
dmvn(x, mu, sigma, log.prob = FALSE)

rmvn(n, mu, sigma)
}
\arguments{
\item{x}{numeric matrix.}

\item{mu}{numeric vector.}

\item{sigma}{numeric matrix.}

\item{log.prob}{logical; if TRUE, probabilities p are given as log(p).}

\item{n}{number of observations. If length(n) > 1,
the length is taken to be the number required.}
}
\description{
Multivariate normal distribution
}
\references{
Nino Hardt, Dicko Ahmadou (Jul 13, 2013).
Faster Multivariate Normal densities with RcppArmadillo and OpenMP.
\url{http://gallery.rcpp.org/articles/dmvnorm_arma/}

Ahmadou Dicko (Mar 12, 2013).
Generating a multivariate gaussian distribution using RcppArmadillo.
\url{http://gallery.rcpp.org/articles/simulate-multivariate-normal/}
}
\author{
Nino Hardt, Dicko Ahmadou, Tymoteusz Wolodzko
}
