% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mvn-arma.R
\name{dmvn}
\alias{dmvn}
\alias{rmvn}
\title{Multivariate normal distribution}
\usage{
dmvn(x, mu, sigma, log.prob = FALSE)

rmvn(n, mu, sigma)
}
\arguments{
\item{x}{n*k numeric matrix.}

\item{mu}{numeric vector of length k.}

\item{sigma}{k*k numeric matrix.}

\item{log.prob}{logical; if \code{TRUE}, probabilities p are given as log(p).}

\item{n}{number of observations. If length(n) > 1,
the length is taken to be the number required.}
}
\description{
Multivariate normal distribution
}
\details{
Multivariate normal probability density function is

\deqn{
f(x_1,\dots,x_n) = \frac{1}{ (2\pi)^{k/2} \sqrt{\mathrm{det}(\Sigma)} }
\exp\left\{ -\frac{1}{2}(\mathbf{x}-\boldsymbol{\mu})' \Sigma^{-1} (\mathbf{x}-\boldsymbol{\mu}) \right\}
}{
f(x) = 1/[(2\pi)^(k/2) * sqrt(det(\Sigma))] * exp( -1/2 (x-\mu)' inv(\Sigma) (x-\mu) )
}

Random generation from this distribution is possible by taking

\deqn{
x = A' z + \mu
}{
x = A' z + \mu
}

where \eqn{z} is a vector of \eqn{k} i.i.d. standard normal deviates,
\eqn{\mu} is a vector of means and \eqn{A} is a \eqn{k \times k}{k*k}
matrix such that \eqn{A'A=\Sigma}{A'A=\Sigma} (\eqn{A} is a Cholesky
factor of \eqn{\Sigma}).

RcppArmadillo implementation of probability density function and random generation
is based on examples from the papers by Nino Hardt and Dicko Ahmadou.
}
\references{
Gentle, J.E. (2006). Random number generation and Monte Carlo methods. Springer.

Nino Hardt, Dicko Ahmadou (Jul 13, 2013).
Faster Multivariate Normal densities with RcppArmadillo and OpenMP.
\url{http://gallery.rcpp.org/articles/dmvnorm_arma/}

Ahmadou Dicko (Mar 12, 2013).
Generating a multivariate gaussian distribution using RcppArmadillo.
\url{http://gallery.rcpp.org/articles/simulate-multivariate-normal/}
}
