% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/kernelboot.R
\name{kernelboot}
\alias{kernelboot}
\title{Kernel density bootstrap}
\usage{
kernelboot(data, statistic, R = 500, bw = c("default", "hns", "silverman",
  "bw.nrd0", "bw.nrd", "bw.ucv", "bw.bcv", "bw.SJ"), kernel = c("mvn",
  "gaussian"), adjust = 1, weights, exclude, ...)
}
\arguments{
\item{data}{Data.}

\item{statistic}{A function which when applied to data returns a vector containing
the statistic(s) of interest. The first argument passed will always
be the original data. Any further arguments can be passed to
\code{statistic} through the \code{...} argument.}

\item{R}{The number of bootstrap replicates.}

\item{bw}{the smoothing bandwidth to be used. The kernels are scaled such that
this is the standard deviation, or covariance matrix of the smoothing kernel.
If provided as numeric value it must be: single numeric value for univariate data;
vector of numbers for multivariate data (length equal to \code{ncol(data)},
not counting the excluded columns declared in \code{exclude});
or matrix of numbers for \code{mvn} kernel (number of columns and rows
equal to \code{ncol(data)}, not counting the excluded columns declared
in \code{exclude}).}

\item{kernel}{a character string giving the smoothing kernel to be used.}

\item{adjust}{the bandwidth used is actually \code{adjust*bw}. This makes it easy to specify
values like ‘half the default’ bandwidth.}

\item{weights}{Vector or matrix of importance weights. It should have as many
elements as there are observations in \code{data}.}

\item{exclude}{a character vector naming columns \emph{not} to be sampled from the kernel
density. Regular bootstrap is applied to those columns.}
}
\description{
Kernel density bootstrap
}

