% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/kernelboot.R
\name{kernelboot}
\alias{kernelboot}
\title{Kernel density bootstrap}
\usage{
kernelboot(data, statistic, R = 500, bw, kernel = c("epanechnikov",
  "gaussian", "rectangular", "triangular", "biweight", "triweight", "cosine",
  "optcosine"), preserve.var = TRUE, adjust = 1, weights = NULL, ...)
}
\arguments{
\item{data}{Data.}

\item{statistic}{A function which when applied to data returns a vector containing
the statistic(s) of interest. The first argument passed will always
be the original data. Any further arguments can be passed to
\code{statistic} through the \code{...} argument.}

\item{R}{The number of bootstrap replicates.}

\item{bw}{the smoothing bandwidth to be used. The kernels are scaled such that
this is the standard deviation, or covariance matrix of the smoothing kernel.
If missing, by default \code{\link[stats]{bw.nrd0}} is used for univariate data,
and \code{\link{bw.silv86}} is used for multivariate data.}

\item{kernel}{a character string giving the smoothing kernel to be used.}

\item{preserve.var}{logical, if \code{TRUE}, then the bootstrap samples preserve sample variance.}

\item{adjust}{the bandwidth used is actually \code{adjust*bw}. This makes it easy to specify
values like 'half the default' bandwidth.}

\item{weights}{Vector of importance weights. It should have as many
elements as there are observations in \code{data}.}

\item{\dots}{further arguments passed to \code{statistic}.}
}
\description{
Kernel density bootstrap
}
\references{
Silverman, B. W. (1986). Density estimation for statistics and data analysis.
Chapman and Hall/CRC.

Wand, M. P. and Jones, M. C. (1995). Kernel Smoothing. Chapman and Hall/CRC.

Scott, D. W. (1992). Multivariate density estimation: theory, practice,
and visualization. John Wiley & Sons.
}
\seealso{
\code{\link{bandwidth}}, \code{\link[stats]{density}},
         \code{\link[stats]{bw.nrd}}, \code{\link[boot]{boot}}
}

